pyproject.toml
[project]
name = "telco-agentic-rules-engine"
version = "0.1.0"
description = "ADK Agentic Rules Engine POC"
authors = [
    {name = "Wilson Dayakar",email = "wilson.dayakar@gmail.com"}
]
readme = "README.md"
requires-python = "^3.14"
dependencies = [
    "google-adk",
    "transitions (>=0.9.3,<0.10.0)",
    "simpleeval (>=1.0.3,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "mcp (>=1.26.0,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
